Moved components to components folder
