+++ src/components/Metropolis/components/icons/CrossIcon.tsx
