+++ src/components/Metropolis/components/icons/LeftArrowIcon.tsx
