+++ src/components/Metropolis/components/icons/PPPIcon.tsx
